Improved example code created by archetype (#19342)
A small change to the example code that hopefully helps new
users to get the idea of Vaadin UI development better. It was
seen in the recent challenge that many newcomers don't
realize that they can access the values of fields directly
(but save the field values to other java fields using value
change listener).

Change-Id: Ifa9c74b35da2e04df25adb54a07ee285d87473de